+++ examples/js-plugin/src/index.js
@@ -11,12 +11,12 @@ let loadedImage = null;
 
 // Configuration
 let options = {
-  modules: ['UPLOAD', 'UPLOADED_IMAGES', 'ICONS_GALLERY', 'IMAGES_GALLERY', 'TAGGING'], // optional default: 'UPLOAD', 'UPLOADED_IMAGES', 'ICONS_GALLERY', 'IMAGES_GALLERY'
+  modules: ['UPLOAD', 'MY_GALLERY', 'ICONS_GALLERY', 'IMAGES_GALLERY', 'TAGGING'], // optional default: 'UPLOAD', 'MY_GALLERY', 'ICONS_GALLERY', 'IMAGES_GALLERY'
   uploadParams: {                 // optional default: {}
     dir: '/dima_test_2'
   },
   // elementID: 'airstore-uploader', // optional default : 'airstore-uploader'
-  folders: [                             // required if UPLOADED_IMAGES is set
+  folders: [                             // required if MY_GALLERY is set
     { dir: '/dima_test_2', label: 'All' },
     //{ dir: '/company_test/project_test', label: 'Project' }
   ],