+++ projects/react-plugin/config.js
@@ -1,5 +1,5 @@
 export default {
-  modules: ['UPLOAD', 'UPLOADED_IMAGES', 'ICONS_GALLERY', 'IMAGES_GALLERY'], // 'UPLOAD', 'UPLOADED_IMAGES', 'ICONS_GALLERY', 'IMAGES_GALLERY'
+  modules: ['UPLOAD', 'MY_GALLERY', 'ICONS_GALLERY', 'IMAGES_GALLERY'],
   folders: [{ dir: '/', label: 'All' }],
   initialTab: 'UPLOAD',
   airstoreUploadKey: '',