+++ Changelog.md
@@ -2,3 +2,7 @@
 
 
 A list of changes to the Airstore Uploader with each new version. Until 1.0.0 is released, breaking changes will be added as minor version bumps since the project is moving quickly.
+
+1. **UPLOADED_IMAGES** module was renamed to **MY_GALLERY**
+
+**Available modules**: **UPLOAD**, **MY_GALLERY**, **ICONS_GALLERY**, **IMAGES_GALLERY**, **TAGGING**